refactor: loadUserByUsername seperate
