+++ src/main/java/com/groom/tennis_match/auth/controller/AdminAuthController.java
@@ -1,6 +1,6 @@
 package com.groom.tennis_match.auth.controller;
 
-import com.groom.tennis_match.auth.AdminAuthService;
+import com.groom.tennis_match.auth.service.AdminAuthService;
 import com.groom.tennis_match.auth.dto.AdminAccountCreateDTO;
 import com.groom.tennis_match.auth.dto.AdminAccountRegisterDTO;
 import com.groom.tennis_match.common.constant.SuccessCode;
