+++ src/main/java/com/groom/tennis_match/auth/service/AdminAuthService.java
@@ -1,5 +1,6 @@
-package com.groom.tennis_match.auth;
+package com.groom.tennis_match.auth.service;
 
+import com.groom.tennis_match.auth.AdminRole;
 import com.groom.tennis_match.auth.dto.AdminAccountCreateDTO;
 import com.groom.tennis_match.auth.dto.AdminAccountRegisterDTO;
 import com.groom.tennis_match.auth.entity.Admin;
@@ -10,7 +11,6 @@ import com.groom.tennis_match.auth.util.PasswordUtil;
 import com.groom.tennis_match.auth.util.SecurityContextUtil;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
@@ -21,27 +21,13 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @RequiredArgsConstructor
 @Transactional
-public class AdminAuthService implements UserDetailsService {
+public class AdminAuthService {
 
     private final AdminRepository adminRepository;
     private final PasswordEncoder passwordEncoder;
     private final SecurityContextUtil securityContextUtil;
     private final PasswordUtil passwordUtil;
 
-    @Override
-    public Admin loadUserByUsername(String username) throws UsernameNotFoundException {
-        log.debug("Spring Security loadUserByUsername 호출: username={}", username);
-
-        Admin user = adminRepository.findByUsername(username)
-                .orElseThrow(() -> {
-                    log.warn("Spring Security - 사용자 없음: username={}", username);
-                    return new UsernameNotFoundException("User not found: " + username);
-                });
-
-        log.debug("Spring Security - 사용자 로드 성공: username={}, authorities={}",
-                username, user.getAuthorities());
-        return user;
-    }
 
     /**
      * 계정 발급 메소드입니다.
